+++ apps/desktop/desktop_native/ssh_agent/src/lib.rs
@@ -5,3 +5,4 @@
 #![allow(dead_code)] // TODO remove when all code is used in follow-up PR
 
 mod crypto;
+mod storage;
