+++ apps/desktop/desktop_native/ssh_agent/src/storage/mod.rs
@@ -1,0 +1,6 @@
+//! Contains all data structures and storage implementations for managing SSH keys during the
+//! Agent's runtime.
+
+pub mod keydata;
+pub mod keystore;
+mod serialization;
